pyproject.toml
[project]
name = "mcp-chainlit-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.2",
]
